requirements.txt
streamlit==1.34.0
google-generativeai==0.7.0
Pillow==11.2.1
python-dotenv==1.0.1
requests==2.32.3
